refactor args
