+++ src/bin/main.rs
@@ -1,63 +1,21 @@
 use actix_web::dev::Server;
 use clap::Parser;
-use log::{error, info, warn};
-use martin::config::{read_config, ConfigBuilder};
-use martin::pg::config::{parse_pg_args, PgArgs, PgConfig};
+use log::info;
+use martin::args::Args;
+use martin::config::{read_config, Config};
+use martin::pg::config::PgConfig;
 use martin::source::IdResolver;
-use martin::srv::config::{SrvArgs, SrvConfig};
 use martin::srv::server;
 use martin::srv::server::RESERVED_KEYWORDS;
 use martin::Error::ConfigWriteError;
 use martin::Result;
-use std::collections::HashMap;
 use std::env;
 use std::ffi::OsStr;
 use std::fmt::Display;
 use std::fs::File;
 use std::io::Write;
-use std::path::PathBuf;
 
 const VERSION: &str = env!("CARGO_PKG_VERSION");
-
-#[derive(Parser, Debug)]
-#[command(about, version)]
-pub struct Args {
-    /// Database connection string
-    pub connection: Vec<String>,
-    /// Path to config file.
-    #[arg(short, long)]
-    pub config: Option<PathBuf>,
-    /// Save resulting config to a file or use "-" to print to stdout.
-    /// By default, only print if sources are auto-detected.
-    #[arg(long)]
-    pub save_config: Option<PathBuf>,
-    /// [Deprecated] Scan for new sources on sources list requests
-    #[arg(short, long, hide = true)]
-    pub watch: bool,
-    #[command(flatten)]
-    srv: SrvArgs,
-    #[command(flatten)]
-    pg: PgArgs,
-}
-
-impl From<Args> for ConfigBuilder {
-    fn from(args: Args) -> Self {
-        if args.watch {
-            warn!("The --watch flag is no longer supported, and will be ignored");
-        }
-        if env::var_os("WATCH_MODE").is_some() {
-            warn!(
-                "The WATCH_MODE environment variable is no longer supported, and will be ignored"
-            );
-        }
-
-        ConfigBuilder {
-            srv: SrvConfig::from(args.srv),
-            postgres: parse_pg_args(&args.pg, &args.connection),
-            unrecognized: HashMap::new(),
-        }
-    }
-}
 
 async fn start(args: Args) -> Result<Server> {
     info!("Starting Martin v{VERSION}");
@@ -70,12 +28,12 @@ async fn start(args: Args) -> Result<Server> {
         info!("Config file is not specified, auto-detecting sources");
         None
     };
-    let mut builder = ConfigBuilder::from(args);
+    let mut args_cfg = Config::try_from(args)?;
     if let Some(file_cfg) = file_cfg {
-        builder.merge(file_cfg);
+        args_cfg.merge(file_cfg);
     }
     let id_resolver = IdResolver::new(RESERVED_KEYWORDS);
-    let mut config = builder.finalize()?;
+    let mut config = args_cfg.finalize()?;
     let sources = config.resolve(id_resolver).await?;
 
     if let Some(file_name) = save_config {
@@ -96,7 +54,7 @@ async fn start(args: Args) -> Result<Server> {
     } else if config
         .postgres
         .iter()
-        .any(|v| v.iter().any(PgConfig::is_autodetect))
+        .any(|v| v.as_slice().iter().any(PgConfig::is_autodetect))
     {
         info!("Martin has been configured with automatic settings.");
         info!("Use --save-config to save or print Martin configuration.");
@@ -122,6 +80,6 @@ async fn main() {
 }
 
 fn on_error<E: Display>(e: E) -> ! {
-    error!("{e}");
+    eprintln!("{e}");
     std::process::exit(1);
 }
