+++ src/lib.rs
@@ -6,11 +6,14 @@
 #![allow(clippy::missing_panics_doc)]
 #![allow(clippy::module_name_repetitions)]
 
+pub mod args;
 pub mod config;
+pub mod one_or_many;
 pub mod pg;
 pub mod source;
 pub mod srv;
 pub mod utils;
+
 pub use crate::utils::Error;
 pub use crate::utils::Result;
