+++ src/pg/config.rs
@@ -1,195 +1,21 @@
-use crate::config::{report_unrecognized_config, set_option, OneOrMany};
+use crate::config::{report_unrecognized_config, set_option};
+use crate::pg::config_function::FuncInfoSources;
+use crate::pg::config_table::TableInfoSources;
 use crate::pg::configurator::PgBuilder;
 use crate::pg::pool::Pool;
-use crate::pg::utils::create_tilejson;
 use crate::pg::utils::PgError::NoConnectionString;
 use crate::pg::utils::Result;
 use crate::source::IdResolver;
 use crate::srv::server::Sources;
-use crate::utils::{get_env_str, InfoMap, Schemas};
+use crate::utils::Schemas;
 use futures::future::try_join;
-use itertools::Itertools;
-use log::warn;
 use serde::{Deserialize, Serialize};
-use serde_yaml::Value;
-use std::collections::{BTreeSet, HashMap};
-use tilejson::{Bounds, TileJSON};
-
-pub const POOL_SIZE_DEFAULT: u32 = 20;
-
-#[derive(clap::Args, Debug)]
-#[command(about, version)]
-pub struct PgArgs {
-    /// Loads trusted root certificates from a file. The file should contain a sequence of PEM-formatted CA certificates.
-    #[cfg(feature = "ssl")]
-    #[arg(long)]
-    pub ca_root_file: Option<std::path::PathBuf>,
-    /// Trust invalid certificates. This introduces significant vulnerabilities, and should only be used as a last resort.
-    #[cfg(feature = "ssl")]
-    #[arg(long)]
-    pub danger_accept_invalid_certs: bool,
-    /// If a spatial table has SRID 0, then this default SRID will be used as a fallback.
-    #[arg(short, long)]
-    pub default_srid: Option<i32>,
-    #[arg(help = format!("Maximum connections pool size [DEFAULT: {}]", POOL_SIZE_DEFAULT), short, long)]
-    pub pool_size: Option<u32>,
-}
-
-#[derive(Clone, Debug, Serialize, Deserialize, PartialEq, Default)]
-pub struct TableInfo {
-    /// Table schema
-    pub schema: String,
-
-    /// Table name
-    pub table: String,
-
-    /// Geometry SRID
-    pub srid: i32,
-
-    /// Geometry column name
-    pub geometry_column: String,
-
-    /// Feature id column name
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub id_column: Option<String>,
-
-    /// An integer specifying the minimum zoom level
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub minzoom: Option<u8>,
-
-    /// An integer specifying the maximum zoom level. MUST be >= minzoom
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub maxzoom: Option<u8>,
-
-    /// The maximum extent of available map tiles. Bounds MUST define an area
-    /// covered by all zoom levels. The bounds are represented in WGS:84
-    /// latitude and longitude values, in the order left, bottom, right, top.
-    /// Values may be integers or floating point numbers.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub bounds: Option<Bounds>,
-
-    /// Tile extent in tile coordinate space
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub extent: Option<u32>,
-
-    /// Buffer distance in tile coordinate space to optionally clip geometries
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub buffer: Option<u32>,
-
-    /// Boolean to control if geometries should be clipped or encoded as is
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub clip_geom: Option<bool>,
-
-    /// Geometry type
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub geometry_type: Option<String>,
-
-    /// List of columns, that should be encoded as tile properties
-    pub properties: HashMap<String, String>,
-
-    /// Mapping of properties to the actual table columns
-    #[serde(skip_deserializing, skip_serializing)]
-    pub prop_mapping: HashMap<String, String>,
-
-    #[serde(flatten, skip_serializing)]
-    pub unrecognized: HashMap<String, Value>,
-}
-
-impl PgInfo for TableInfo {
-    fn format_id(&self) -> String {
-        format!("{}.{}.{}", self.schema, self.table, self.geometry_column)
-    }
-
-    fn to_tilejson(&self) -> TileJSON {
-        create_tilejson(
-            self.format_id(),
-            self.minzoom,
-            self.maxzoom,
-            self.bounds,
-            None,
-        )
-    }
-}
+use tilejson::TileJSON;
 
 pub trait PgInfo {
     fn format_id(&self) -> String;
     fn to_tilejson(&self) -> TileJSON;
 }
-
-#[derive(Clone, Debug, Serialize, Deserialize, PartialEq, Default)]
-pub struct FunctionInfo {
-    /// Schema name
-    pub schema: String,
-
-    /// Function name
-    pub function: String,
-
-    /// An integer specifying the minimum zoom level
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub minzoom: Option<u8>,
-
-    /// An integer specifying the maximum zoom level. MUST be >= minzoom
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub maxzoom: Option<u8>,
-
-    /// The maximum extent of available map tiles. Bounds MUST define an area
-    /// covered by all zoom levels. The bounds are represented in WGS:84
-    /// latitude and longitude values, in the order left, bottom, right, top.
-    /// Values may be integers or floating point numbers.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub bounds: Option<Bounds>,
-
-    #[serde(flatten, skip_serializing)]
-    pub unrecognized: HashMap<String, Value>,
-}
-
-impl FunctionInfo {
-    #[must_use]
-    pub fn new(schema: String, function: String) -> Self {
-        Self {
-            schema,
-            function,
-            ..Default::default()
-        }
-    }
-
-    #[must_use]
-    pub fn new_extended(
-        schema: String,
-        function: String,
-        minzoom: u8,
-        maxzoom: u8,
-        bounds: Bounds,
-    ) -> Self {
-        Self {
-            schema,
-            function,
-            minzoom: Some(minzoom),
-            maxzoom: Some(maxzoom),
-            bounds: Some(bounds),
-            ..Default::default()
-        }
-    }
-}
-
-impl PgInfo for FunctionInfo {
-    fn format_id(&self) -> String {
-        format!("{}.{}", self.schema, self.function)
-    }
-
-    fn to_tilejson(&self) -> TileJSON {
-        create_tilejson(
-            self.format_id(),
-            self.minzoom,
-            self.maxzoom,
-            self.bounds,
-            None,
-        )
-    }
-}
-
-pub type TableInfoSources = InfoMap<TableInfo>;
-pub type FuncInfoSources = InfoMap<FunctionInfo>;
 
 #[derive(Clone, Debug, Default, PartialEq, Serialize, Deserialize)]
 pub struct PgConfig {
@@ -272,61 +98,7 @@ impl PgConfig {
 }
 
 #[must_use]
-pub fn parse_pg_args(args: &PgArgs, cli_strings: &[String]) -> Option<OneOrMany<PgConfig>> {
-    let mut strings = cli_strings
-        .iter()
-        .filter(|s| is_postgresql_string(s))
-        .map(|s| Some(s.to_string()))
-        .unique()
-        .collect::<BTreeSet<_>>();
-
-    if let Some(s) = get_env_str("DATABASE_URL") {
-        if is_postgresql_string(&s) {
-            strings.insert(Some(s));
-        } else {
-            warn!("Environment variable DATABASE_URL is not a postgres connection string");
-        }
-    }
-
-    if strings.is_empty() {
-        // If there are no connection strings in the CLI, try to parse env vars into a single connection
-        strings.insert(None);
-    }
-
-    let builders: Vec<_> = strings
-        .into_iter()
-        .map(|s| PgConfig {
-            connection_string: s,
-            #[cfg(feature = "ssl")]
-            ca_root_file: args
-                .ca_root_file
-                .clone()
-                .or_else(|| std::env::var_os("CA_ROOT_FILE").map(std::path::PathBuf::from)),
-            #[cfg(feature = "ssl")]
-            danger_accept_invalid_certs: args.danger_accept_invalid_certs
-                || get_env_str("DANGER_ACCEPT_INVALID_CERTS").is_some(),
-            default_srid: args.default_srid.or_else(|| {
-                get_env_str("DEFAULT_SRID").and_then(|srid| match srid.parse::<i32>() {
-                    Ok(v) => Some(v),
-                    Err(v) => {
-                        warn!("Env var DEFAULT_SRID is not a valid integer {srid}: {v}");
-                        None
-                    }
-                })
-            }),
-            pool_size: args.pool_size,
-            ..Default::default()
-        })
-        .collect();
-
-    match builders.len() {
-        0 => None,
-        1 => Some(OneOrMany::One(builders.into_iter().next().unwrap())),
-        _ => Some(OneOrMany::Many(builders)),
-    }
-}
-
-fn is_postgresql_string(s: &str) -> bool {
+pub fn is_postgresql_string(s: &str) -> bool {
     s.starts_with("postgresql://") || s.starts_with("postgres://")
 }
 
@@ -334,8 +106,13 @@ fn is_postgresql_string(s: &str) -> bool {
 mod tests {
     use super::*;
     use crate::config::Config;
+    use crate::one_or_many::OneOrMany::{Many, One};
+    use crate::pg::config_function::FunctionInfo;
+    use crate::pg::config_table::TableInfo;
     use crate::pg::utils::tests::{assert_config, some_str};
     use indoc::indoc;
+    use std::collections::HashMap;
+    use tilejson::Bounds;
 
     #[test]
     #[allow(clippy::too_many_lines)]
@@ -347,11 +124,11 @@ mod tests {
               connection_string: 'postgresql://postgres@localhost/db'
         "},
             &Config {
-                postgres: Some(vec![PgConfig {
+                postgres: Some(One(PgConfig {
                     connection_string: some_str("postgresql://postgres@localhost/db"),
                     run_autodiscovery: true,
                     ..Default::default()
-                }]),
+                })),
                 ..Default::default()
             },
         );
@@ -364,7 +141,7 @@ mod tests {
               - connection_string: 'postgresql://postgres@localhost:5433/db'
         "},
             &Config {
-                postgres: Some(vec![
+                postgres: Some(Many(vec![
                     PgConfig {
                         connection_string: some_str("postgres://postgres@localhost:5432/db"),
                         run_autodiscovery: true,
@@ -375,7 +152,7 @@ mod tests {
                         run_autodiscovery: true,
                         ..Default::default()
                     },
-                ]),
+                ])),
                 ..Default::default()
             },
         );
@@ -414,7 +191,7 @@ mod tests {
                   bounds: [-180.0, -90.0, 180.0, 90.0]
         "},
             &Config {
-                postgres: Some(vec![PgConfig {
+                postgres: Some(One(PgConfig {
                     connection_string: some_str("postgres://postgres@localhost:5432/db"),
                     default_srid: Some(4326),
                     pool_size: Some(20),
@@ -447,7 +224,7 @@ mod tests {
                         ),
                     )])),
                     ..Default::default()
-                }]),
+                })),
                 ..Default::default()
             },
         );
