+++ src/pg/config_function.rs
@@ -1,0 +1,81 @@
+use crate::pg::config::PgInfo;
+use crate::pg::utils::create_tilejson;
+use crate::utils::InfoMap;
+use serde::{Deserialize, Serialize};
+use serde_yaml::Value;
+use std::collections::HashMap;
+use tilejson::{Bounds, TileJSON};
+
+pub type FuncInfoSources = InfoMap<FunctionInfo>;
+
+#[derive(Clone, Debug, Serialize, Deserialize, PartialEq, Default)]
+pub struct FunctionInfo {
+    /// Schema name
+    pub schema: String,
+
+    /// Function name
+    pub function: String,
+
+    /// An integer specifying the minimum zoom level
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub minzoom: Option<u8>,
+
+    /// An integer specifying the maximum zoom level. MUST be >= minzoom
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub maxzoom: Option<u8>,
+
+    /// The maximum extent of available map tiles. Bounds MUST define an area
+    /// covered by all zoom levels. The bounds are represented in WGS:84
+    /// latitude and longitude values, in the order left, bottom, right, top.
+    /// Values may be integers or floating point numbers.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub bounds: Option<Bounds>,
+
+    #[serde(flatten, skip_serializing)]
+    pub unrecognized: HashMap<String, Value>,
+}
+
+impl FunctionInfo {
+    #[must_use]
+    pub fn new(schema: String, function: String) -> Self {
+        Self {
+            schema,
+            function,
+            ..Default::default()
+        }
+    }
+
+    #[must_use]
+    pub fn new_extended(
+        schema: String,
+        function: String,
+        minzoom: u8,
+        maxzoom: u8,
+        bounds: Bounds,
+    ) -> Self {
+        Self {
+            schema,
+            function,
+            minzoom: Some(minzoom),
+            maxzoom: Some(maxzoom),
+            bounds: Some(bounds),
+            ..Default::default()
+        }
+    }
+}
+
+impl PgInfo for FunctionInfo {
+    fn format_id(&self) -> String {
+        format!("{}.{}", self.schema, self.function)
+    }
+
+    fn to_tilejson(&self) -> TileJSON {
+        create_tilejson(
+            self.format_id(),
+            self.minzoom,
+            self.maxzoom,
+            self.bounds,
+            None,
+        )
+    }
+}
