+++ src/pg/utils.rs
@@ -90,10 +90,10 @@ pub fn is_valid_zoom(zoom: i32, minzoom: Option<u8>, maxzoom: Option<u8>) -> boo
 
 #[cfg(test)]
 pub(crate) mod tests {
-    use crate::config::{Config, ConfigBuilder};
+    use crate::config::Config;
 
     pub fn assert_config(yaml: &str, expected: &Config) {
-        let config: ConfigBuilder = serde_yaml::from_str(yaml).expect("parse yaml");
+        let config: Config = serde_yaml::from_str(yaml).expect("parse yaml");
         let actual = config.finalize().expect("finalize");
         assert_eq!(&actual, expected);
     }
