+++ src/utils.rs
@@ -11,6 +11,9 @@ pub type InfoMap<T> = HashMap<String, T>;
 
 #[derive(thiserror::Error, Debug)]
 pub enum Error {
+    #[error("The --config and the connection parameters cannot be used together")]
+    ConfigAndConnectionsError,
+
     #[error("Unable to load config file {}: {0}", .1.display())]
     ConfigLoadError(io::Error, PathBuf),
